conveyor_datamatrix_system/requirements.txt
opencv-python-headless>=4.8.0
pyzbar>=0.1.9
numpy>=1.24.0
PyYAML>=6.0
pillow>=10.0.0
scikit-image>=0.21.0

# Industrial camera SDKs (uncomment based on your camera brand)
# pypylon>=6.0.0  # Basler cameras
# flir-spinnaker>=2.0.0  # FLIR cameras
# harvesters>=3.0.0  # Generic GigE Vision/USB3 Vision

# Optional: GPU acceleration
# cupy-cuda11x>=12.0.0  # For NVIDIA GPUs

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
